--- Source/WeaponProficiencyCompat/WeaponProficiencyCompat/HarmonyStarter.cs
@@ -1,15 +1,0 @@
-﻿using HarmonyLib;
-using Verse;
-
-namespace CombatExtended.Compatibility.WeaponProficiency.HarmonyPatches
-{
-    [StaticConstructorOnStartup]
-    public static class HarmonyStarter
-    {
-        static HarmonyStarter()
-        {
-            Harmony harmony = new Harmony("CombatExtended.Compatibility.WeaponProficiency");
-            harmony.PatchAll();
-        }
-    }
-}
